Move test logic out of helper file
